Relocate per-estimator configs into helper packages (batch 2/4)
Move 10 more live per-estimator config classes out of config_models.py into
their estimator's utils/<name>_helpers/config.py, re-exported via the lazy
__getattr__ shim so existing imports keep working:

  RMSI, SPOTSYNTH, SNN, CTSC, ISCM, SPILLSYNTH, DSCAR, PROXIMAL, SCMO, SI

No behavior change. config_models.py: 3468 -> 2649 lines.
