+++ mlsynth/config_models.py
@@ -350,662 +350,9 @@ class BaseEstimatorConfig(BaseModel):
 
 
 
-class RMSIConfig(BaseEstimatorConfig):
-    """Configuration for the RMSI estimator.
-
-    Agarwal, Choi & Yuan (2026), *"Robust Matrix Estimation with Side
-    Information"* (arXiv:2603.24833). Imputes the treated counterfactual of a
-    block-adoption causal panel by a four-component sieve + nuclear-norm matrix
-    estimator that exploits unit-level (row) and time-level (column) covariates.
-    Inherits the standard ``df`` / ``outcome`` / ``treat`` / ``unitid`` /
-    ``time`` interface.
-
-    Parameters
-    ----------
-    unit_covariates : list of str
-        Columns that vary across units and are (approximately) constant within a
-        unit; averaged per unit to form the row feature matrix ``X``. May be
-        empty (then the row projection captures only the mean).
-    time_covariates : list of str
-        Columns that vary across periods and are constant across units within a
-        period; averaged per period to form the column feature matrix ``Z``.
-        May be empty.
-    sieve_order : int
-        Polynomial sieve order ``J`` (default 2, matching the paper).
-    rank : int, optional
-        Factor rank for the block recombination; chosen by a relative-magnitude
-        singular-value threshold when None.
-    """
-
-    unit_covariates: List[str] = Field(
-        default_factory=list,
-        description="Unit-level covariate columns (row side information).")
-    time_covariates: List[str] = Field(
-        default_factory=list,
-        description="Time-level covariate columns (column side information).")
-    sieve_order: int = Field(
-        default=2, ge=1, description="Polynomial sieve order J.")
-    rank: Optional[int] = Field(
-        default=None, ge=1,
-        description="Factor rank; relative-threshold estimate when None.")
-
-
-class SPOTSYNTHConfig(BaseEstimatorConfig):
-    """Configuration for the SPOTSYNTH estimator.
-
-    O'Riordan & Gilligan-Lee (2025), *"Spillover detection for donor selection in
-    synthetic control models"* (Journal of Causal Inference 13:20240036,
-    doi:10.1515/jci-2024-0036). Screens each candidate donor for spillover
-    contamination via a pre-intervention forecast test (Algorithm 1), excludes
-    the contaminated donors, and fits a simplex synthetic control on the valid
-    set. Inherits the standard ``df`` / ``outcome`` / ``treat`` / ``unitid`` /
-    ``time`` interface and expects a single treated unit.
-
-    Parameters
-    ----------
-    selection : {"S1", "S2", "all"}
-        Donor-selection rule. ``S1`` keeps the ``n_donors`` donors with the
-        smallest forecast error (the analyst fixes how many to keep); ``S2``
-        keeps donors whose realised post-intervention value falls inside the
-        ``ppi`` posterior predictive interval (controls the false-positive rate);
-        ``all`` keeps every donor (the unscreened baseline).
-    forecast : {"loo", "lag"}
-        Forecast anchor for the screen. ``loo`` (default) is the leave-one-out
-        anchor: each donor's whole post-intervention trajectory is forecast from
-        the *other* donors' common factors and scored by the mean absolute
-        deviation. It is robust to the *onset speed* of the spillover (it detects
-        gradually-arriving contamination that the first-post-point misses) and is
-        the right choice whenever the donor pool is mostly valid -- the typical
-        applied setting. ``lag`` is the paper's literal Algorithm 1: forecast the
-        *first* post-intervention point from the last clean pre-intervention
-        cross-section. It is the correct anchor only in the paper's stress regime
-        -- a *mostly-invalid* donor pool with a *sharp* (immediate) spillover --
-        where ``loo`` inverts because the contaminated majority defines the
-        "consensus". See the forecast-anchor discussion in the estimator docs.
-    n_donors : int, optional
-        Number of donors to keep under ``S1`` (default: half the pool).
-    ppi : float
-        Posterior-predictive-interval level for ``S2`` (default 0.8).
-    n_factors : int
-        Number of donor factors used to regularise the forecast (default 5).
-    time_average : int, optional
-        Bucket width for time-averaging the data before screening (``lag`` only;
-        reduces false negatives with very noisy donors).
-    inference : {"bayes", "frequentist"}
-        Synthetic-control weight model. ``bayes`` (default) is the authors'
-        Bayesian simplex SC -- weights with a ``Dirichlet(dirichlet_alpha)``
-        prior, a half-normal prior on the residual sd, pre-period standardisation
-        of target and donors, and 95% posterior-predictive credible intervals --
-        fit with NumPyro's NUTS (the same Hamiltonian-Monte-Carlo family as the
-        authors' Stan). It requires the optional ``numpyro`` package.
-        ``frequentist`` is a fast, dependency-free simplex least-squares point
-        estimate (no intervals), useful for large simulations or when NumPyro is
-        unavailable.
-    dirichlet_alpha : float
-        Dirichlet concentration on the donor weights (paper uses 0.4; ``< 1``
-        favours sparse weights).
-    ci_level : float
-        Credible-interval level for the Bayesian fit (paper uses 0.95).
-    n_samples, n_warmup : int
-        Posterior draws and warm-up iterations for the Bayesian sampler.
-    debias : bool
-        If True, also compute the proximal (two-stage / GMM) debiased ATT
-        (equation 5), using the screen-excluded donors as proximal controls to
-        correct errors-in-variables bias when the kept donors are noisy proxies.
-    seed : int
-        RNG seed for the Bayesian sampler.
-    """
-
-    selection: Literal["S1", "S2", "all"] = Field(
-        default="S1", description="Donor-selection rule (S1, S2, or all).")
-    forecast: Literal["loo", "lag"] = Field(
-        default="loo",
-        description="Forecast anchor: loo (default, onset-robust) or lag "
-                    "(paper Algorithm 1, for mostly-invalid sharp-spillover pools).")
-    n_donors: Optional[int] = Field(
-        default=None, ge=1, description="Number of donors to keep under S1.")
-    ppi: float = Field(
-        default=0.8, gt=0.0, lt=1.0,
-        description="Posterior predictive interval level for S2.")
-    n_factors: int = Field(
-        default=5, ge=1, description="Number of donor factors in the forecast.")
-    time_average: Optional[int] = Field(
-        default=None, ge=1, description="Bucket width for time-averaging (lag).")
-    inference: Literal["bayes", "frequentist"] = Field(
-        default="bayes",
-        description="SC weight model (bayes = NumPyro NUTS, or frequentist simplex LS).")
-    dirichlet_alpha: float = Field(
-        default=0.4, gt=0.0, description="Dirichlet concentration on donor weights.")
-    ci_level: float = Field(
-        default=0.95, gt=0.0, lt=1.0, description="Credible-interval level.")
-    n_samples: int = Field(
-        default=4000, ge=200, description="Posterior draws for the Bayesian fit.")
-    n_warmup: int = Field(
-        default=2000, ge=100, description="Warm-up iterations for the Bayesian fit.")
-    debias: bool = Field(
-        default=False, description="Compute the proximal (GMM) debiased ATT.")
-    seed: int = Field(
-        default=0, description="RNG seed for the Bayesian sampler.")
-
-
-class SNNConfig(BaseEstimatorConfig):
-    """Configuration for the Synthetic Nearest Neighbors (SNN) estimator.
-
-    Agarwal, Dahleh, Shah & Shen (2021), *"Causal Matrix Completion"*
-    (arXiv:2109.15154). Imputes treated units' untreated potential
-    outcomes by MNAR matrix completion (anchor submatrix + principal
-    component regression), generalising the Synthetic Interventions /
-    synthetic-control approach. Inherits the standard ``df`` / ``outcome``
-    / ``treat`` / ``unitid`` / ``time`` interface.
-
-    Parameters
-    ----------
-    n_neighbors : int
-        Number of synthetic neighbours (anchor-row groups) to average.
-    max_rank : int, optional
-        Fixed PCR truncation rank; overrides the spectral/universal rule.
-    spectral_energy : float
-        Singular-value energy threshold for spectral rank selection
-        (used when ``max_rank`` is None and ``universal_rank`` is False).
-    universal_rank : bool
-        Use the Donoho-Gavish (2014) universal hard-threshold rank.
-        Default True -- well-calibrated for small low-rank panels (e.g.
-        Prop 99); set False to use the spectral-energy threshold.
-    clip : bool
-        Clip imputations to the observed value range.
-    inference : bool
-        Run a leave-one-control jackknife for the ATT SE / CI.
-    alpha : float
-        Two-sided level for the jackknife confidence interval.
-    random_state : int
-        Seed for anchor-row splitting.
-    """
-
-    n_neighbors: int = Field(
-        default=1, ge=1,
-        description="Number of synthetic neighbours (anchor-row groups).",
-    )
-    max_rank: Optional[int] = Field(
-        default=None, ge=1,
-        description="Fixed PCR truncation rank (overrides spectral rule).",
-    )
-    spectral_energy: float = Field(
-        default=0.95, gt=0.0, le=1.0,
-        description="Singular-value energy threshold for rank selection.",
-    )
-    universal_rank: bool = Field(
-        default=True,
-        description="Use the Donoho-Gavish universal hard-threshold rank "
-                    "(default; well-calibrated for small low-rank panels). "
-                    "Set False to use the spectral-energy threshold instead.",
-    )
-    clip: bool = Field(
-        default=True,
-        description="Clip imputations to the observed value range.",
-    )
-    inference: bool = Field(
-        default=False,
-        description="Run a leave-one-control jackknife for the ATT SE/CI.",
-    )
-    alpha: float = Field(
-        default=0.05, gt=0.0, lt=1.0,
-        description="Two-sided level for the jackknife confidence interval.",
-    )
-    random_state: int = Field(
-        default=0,
-        description="Seed for anchor-row splitting.",
-    )
-
-
-class CTSCConfig(BaseEstimatorConfig):
-    """Configuration for the Continuous-Treatment Synthetic Control (CTSC).
-
-    Powell, D. (2022). *"Synthetic Control Estimation Beyond Comparative
-    Case Studies,"* Journal of Business & Economic Statistics. Generalises
-    synthetic control to continuous / multi-valued treatments with no clean
-    treated/never-treated split; jointly estimates unit-specific treatment
-    slopes and synthetic controls for all units. (The paper calls it "GSC";
-    mlsynth uses CTSC to avoid collision with Xu (2017)'s GSC.)
-
-    Parameters
-    ----------
-    treatment_vars : list of str
-        The ``K >= 1`` treatment / explanatory columns (continuous or
-        discrete). CTSC estimates an average marginal effect for each.
-    population_col : str, optional
-        Time-invariant per-unit weight column for the average effect
-        (e.g. population). Defaults to uniform weights.
-    use_fit_weights : bool
-        Use the two-step per-unit fit weights ``Omega_i`` (paper eq. 6).
-        Default True.
-    inference : bool
-        Run the sign-flip Wald test of ``H0: alpha^AE = 0``. Default True.
-    n_draws : int
-        Rademacher draws for the randomization test.
-    random_state : int
-        Seed for the randomization-test RNG.
-
-    Notes
-    -----
-    The base ``treat`` field is unused by CTSC; provide the continuous /
-    discrete treatment column(s) via ``treatment_vars`` instead. Pass any
-    existing column name for ``treat`` to satisfy the base config.
-    """
-
-    treatment_vars: List[str] = Field(
-        ...,
-        description="The K >= 1 continuous/discrete treatment columns.",
-    )
-    population_col: Optional[str] = Field(
-        default=None,
-        description="Per-unit weight column for the average effect "
-                    "(default uniform).",
-    )
-    use_fit_weights: bool = Field(
-        default=True,
-        description="Use the two-step per-unit fit weights Omega_i.",
-    )
-    inference: bool = Field(
-        default=True,
-        description="Run the sign-flip Wald test of H0: alpha^AE = 0.",
-    )
-    n_draws: int = Field(
-        default=2000, ge=100,
-        description="Rademacher draws for the randomization test.",
-    )
-    random_state: int = Field(
-        default=0,
-        description="Seed for the randomization-test RNG.",
-    )
-
-
-class ISCMConfig(BaseEstimatorConfig):
-    """Configuration for the Imperfect Synthetic Controls (ISCM) estimator.
-
-    Powell, D. (2026). *"Imperfect Synthetic Controls,"* Journal of Applied
-    Econometrics. Builds synthetic controls for every unit, identifies the
-    treatment effect even when the treated unit is outside the convex hull,
-    weights units by a data-driven fit metric, and uses Ibragimov-Muller
-    inference valid for small donor pools. Inherits the standard ``df`` /
-    ``outcome`` / ``treat`` / ``unitid`` / ``time`` interface.
-
-    Parameters
-    ----------
-    inference : bool
-        Run Ibragimov-Muller inference over the per-unit estimates.
-        Default True.
-    null_value : float
-        Null effect ``alpha_0`` for the randomization test. Default 0.
-    alpha : float
-        Two-sided level for the confidence interval.
-    n_draws : int
-        Number of Rademacher sign-flip draws for the p-value.
-    random_state : int
-        Seed for the randomization-test RNG.
-    """
-
-    inference: bool = Field(
-        default=True,
-        description="Run Ibragimov-Muller inference over per-unit estimates.",
-    )
-    null_value: float = Field(
-        default=0.0,
-        description="Null effect alpha_0 for the randomization test.",
-    )
-    alpha: float = Field(
-        default=0.05, gt=0.0, lt=1.0,
-        description="Two-sided level for the confidence interval.",
-    )
-    n_draws: int = Field(
-        default=10000, ge=100,
-        description="Number of Rademacher sign-flip draws for the p-value.",
-    )
-    random_state: int = Field(
-        default=0,
-        description="Seed for the randomization-test RNG.",
-    )
-
-
 # VanillaSCConfig has been relocated to mlsynth/utils/vanillasc_helpers/config.py
 # (co-located with the estimator's helpers). It is re-exported from this
 # module for backward compatibility via the module-level __getattr__ below.
-
-
-class SPILLSYNTHConfig(BaseEstimatorConfig):
-    """Configuration for the SPILLSYNTH estimator.
-
-    Spillover-aware synthetic-control wrapper. Currently ships the
-    Cao & Dowd (2023) estimator under ``method='cd'``: a Ferman-Pinto
-    demeaned SCM is fit leave-one-out for every unit, and the
-    treatment-and-spillover effect vector is recovered jointly under
-    a user-specified spillover structure (Example 3 of the paper:
-    each declared affected unit gets its own free spillover
-    coefficient). Additional methods (e.g. iSCM, distance-decay
-    spillover, P-test inference) will be added behind the same
-    ``method`` dispatcher.
-
-    Parameters
-    ----------
-    method : {"cd"}
-        Estimation method. Only ``"cd"`` (Cao-Dowd 2023) is
-        implemented for now.
-    affected_units : list, optional
-        Labels (matching values of ``unitid``) of control units the
-        researcher believes are potentially exposed to spillover from
-        the treated unit. Each gets its own free spillover coefficient
-        in the resulting estimator. If ``None`` (the default), no
-        affected units are declared and the estimator reduces to
-        vanilla demeaned SCM. The treated unit must NOT appear in this
-        list.
-    solver : str, optional
-        ``cvxpy`` solver name used for the leave-one-out simplex SCM
-        fits. Defaults to ``"CLARABEL"``.
-
-    References
-    ----------
-    Cao, J., & Dowd, C. (2023). *Estimation and Inference for
-    Synthetic Control Methods with Spillover Effects.* Working paper.
-
-    Ferman, B., & Pinto, C. (2021). *Synthetic Controls with Imperfect
-    Pretreatment Fit.* Quantitative Economics, 12(4), 1197-1221.
-    """
-
-    method: Literal["cd", "iscm", "grossi", "sar"] = Field(
-        default="cd",
-        description="Spillover-aware SCM method. 'cd' = Cao & Dowd (2023); "
-                    "'iscm' = inclusive SCM (Di Stefano & Mellace 2024); "
-                    "'grossi' = Grossi et al. (2025) direct + spillover effects "
-                    "under partial interference (penalized SC on far controls); "
-                    "'sar' = Sakaguchi & Tagawa (2026) spatial-autoregressive "
-                    "Bayesian SCM (relaxes SUTVA via a SAR model on the control "
-                    "outcomes; needs spatial_W/spatial_w). "
-                    "For 'grossi', affected_units lists the treated unit's "
-                    "cluster-mates (excluded from the donor pool, given "
-                    "spillover effects).",
-    )
-    spatial_W: Optional[Any] = Field(
-        default=None,
-        description="(method='sar') Control-to-control spatial-weight matrix. "
-                    "A labelled (N x N) pandas DataFrame indexed by control unit "
-                    "label (aligned automatically), or a bare N x N array in "
-                    "control-label order. Row-normalised internally.",
-    )
-    spatial_w: Optional[Any] = Field(
-        default=None,
-        description="(method='sar') Treated-to-control spatial-weight vector "
-                    "(length N): a pandas Series/DataFrame keyed by unit label, "
-                    "a dict, or a bare array in control-label order. Normalised "
-                    "to sum to one internally.",
-    )
-    p_factors: int = Field(
-        default=1, ge=0,
-        description="(method='sar') Number of AR(1) latent factors in the SAR "
-                    "panel model. 0 disables the factor block.",
-    )
-    mcmc_iter: int = Field(
-        default=6000, ge=2,
-        description="(method='sar') Total MCMC iterations per step.",
-    )
-    mcmc_burn: int = Field(
-        default=2000, ge=1,
-        description="(method='sar') Burn-in iterations per step.",
-    )
-    step_rho: float = Field(
-        default=0.02, gt=0.0,
-        description="(method='sar') Random-walk Metropolis step for rho.",
-    )
-    mcmc_seed: int = Field(
-        default=0, description="(method='sar') RNG seed for the sampler.",
-    )
-    propagate_alpha: bool = Field(
-        default=True,
-        description="(method='sar') If True (default) the credible bands pair "
-                    "each rho draw with a synthetic-weight (alpha) draw, "
-                    "propagating posterior uncertainty in alpha -- the "
-                    "statistically correct interval that attains nominal "
-                    "coverage in simulation. If False, alpha is held at its "
-                    "posterior mean when sweeping rho (narrower bands; the "
-                    "authors' empirical convention). Point estimates are "
-                    "unaffected.",
-    )
-    covariates: Optional[List[str]] = Field(
-        default=None,
-        description="(method='iscm' only) Covariate columns to match on, "
-                    "averaged over the pre-treatment period and fed to the "
-                    "bilevel predictor-matching solver. When None, the "
-                    "inclusive SCM matches on pre-treatment outcomes only.",
-    )
-    bilevel_solver: Literal["malo", "mscmt", "penalized"] = Field(
-        default="malo",
-        description="(method='iscm' with covariates) Bilevel backend for "
-                    "predictor matching: 'malo' (Malo et al. 2024 corner "
-                    "search), 'mscmt' (Becker-Kloessner 2018 global "
-                    "differential-evolution search), or 'penalized' "
-                    "(Abadie-L'Hour 2021 pairwise-penalized estimator with "
-                    "leave-out lambda selection -- a unique, sparse solution).",
-    )
-    covariate_windows: Optional[Dict[Any, Any]] = Field(
-        default=None,
-        description="(method='iscm'/'grossi') Per-covariate averaging window as "
-                    "an inclusive (start, end) range of time labels, e.g. "
-                    "{'invest': (1964, 1969), 'popdens': (1969, 1969)} "
-                    "(Abadie's special-predictor spec). Covariates not listed "
-                    "are averaged over the full pre-treatment period.",
-    )
-    bias_correct: bool = Field(
-        default=False,
-        description="(method='iscm'/'grossi') Apply the Abadie-L'Hour (2021) "
-                    "bias correction to each unit's synthetic-control gap, "
-                    "removing the part attributable to residual covariate "
-                    "imbalance via a ridge regression of the outcome on the "
-                    "covariates. Requires 'covariates'; most useful when the "
-                    "covariates genuinely explain the outcome.",
-    )
-    n_boot: int = Field(
-        default=0, ge=0,
-        description="(method='grossi') Residual-resampling draws for the "
-                    "pivotal bias-corrected confidence intervals (Grossi et "
-                    "al. eqs. 3.6-3.7). 0 (default) skips inference.",
-    )
-    ci_level: float = Field(
-        default=0.90, gt=0.0, lt=1.0,
-        description="(method='grossi') Confidence level for the residual-"
-                    "resampling intervals. The paper uses 0.90.",
-    )
-    seed: int = Field(
-        default=0,
-        description="(method='grossi') RNG seed for residual resampling.",
-    )
-    affected_units: Optional[List[Any]] = Field(
-        default=None,
-        description=(
-            "Labels of control units potentially exposed to spillover. "
-            "Required for spillover_structure='per_unit' (when p > 0) "
-            "and 'homogeneous'. Ignored for 'distance_decay' (which "
-            "uses unit_distances). The treated unit must NOT appear here."
-        ),
-    )
-    spillover_structure: Literal["per_unit", "homogeneous", "distance_decay"] = Field(
-        default="per_unit",
-        description=(
-            "A-matrix construction (Cao-Dowd v3 Examples 1/2/3): "
-            "'per_unit' (Example 1, default, each affected unit gets its "
-            "own free coefficient), 'homogeneous' (Example 2, shared "
-            "coefficient b across affected units), 'distance_decay' "
-            "(Example 3, alpha_i = b * exp(-d_i) via unit_distances)."
-        ),
-    )
-    unit_distances: Optional[Dict[Any, float]] = Field(
-        default=None,
-        description=(
-            "Required when spillover_structure='distance_decay'. Maps "
-            "unit label to a non-negative scalar distance from the "
-            "treated unit. Controls absent from the dict are treated as "
-            "infinitely far (zero decay weight)."
-        ),
-    )
-    weighting: Literal["identity", "efficient"] = Field(
-        default="identity",
-        description=(
-            "Weighting matrix in the SP estimator. 'identity' = the "
-            "standard W = I estimator. 'efficient' = additionally "
-            "compute the GMM-weighted variant of Cao-Dowd v3 "
-            "Proposition S.1 using W = sample-Omega^{-1} (lower "
-            "asymptotic variance); exposed via results.cd.efficient_fit."
-        ),
-    )
-    solver: Optional[str] = Field(
-        default=None,
-        description="cvxpy solver name for the leave-one-out SCM fits. "
-                    "Defaults to CLARABEL.",
-    )
-
-    @model_validator(mode="after")
-    def _check_affected_units(cls, values: Any) -> Any:
-        au = values.affected_units
-        structure = values.spillover_structure
-        distances = values.unit_distances
-
-        # Structure-specific requirements.
-        if structure == "homogeneous" and (au is None or len(au) == 0):
-            raise MlsynthConfigError(
-                "SPILLSYNTH: spillover_structure='homogeneous' needs at "
-                "least one entry in affected_units."
-            )
-        if structure == "distance_decay":
-            if not isinstance(distances, dict) or len(distances) == 0:
-                raise MlsynthConfigError(
-                    "SPILLSYNTH: spillover_structure='distance_decay' "
-                    "requires a non-empty unit_distances={label: d, ...}."
-                )
-
-        if au is None:
-            return values
-        if len(set(au)) != len(au):
-            raise MlsynthConfigError(
-                "SPILLSYNTH: affected_units contains duplicate labels."
-            )
-        # Resolve treated units on the spot to surface helpful errors
-        # before .fit() runs. Multiple treated units are allowed (Cao-
-        # Dowd v3 Section S.1.2), but none of them may also appear in
-        # affected_units.
-        df = values.df
-        treat_col = values.treat
-        unit_col = values.unitid
-        treated_rows = set(df.loc[df[treat_col] != 0, unit_col].unique())
-        overlap = treated_rows.intersection(au)
-        if overlap:
-            raise MlsynthConfigError(
-                f"SPILLSYNTH: treated units {sorted(overlap, key=str)} cannot "
-                "also appear in affected_units."
-            )
-        present = set(df[unit_col].unique())
-        missing = [u for u in au if u not in present]
-        if missing:
-            raise MlsynthConfigError(
-                f"SPILLSYNTH: affected_units {missing} not in df[{unit_col!r}]."
-            )
-        return values
-
-    @model_validator(mode="after")
-    def _check_sar(cls, values: Any) -> Any:
-        if values.method != "sar":
-            return values
-        if values.spatial_W is None or values.spatial_w is None:
-            raise MlsynthConfigError(
-                "SPILLSYNTH: method='sar' requires both spatial_W (control-to-"
-                "control weight matrix) and spatial_w (treated-to-control weight "
-                "vector)."
-            )
-        if values.mcmc_burn >= values.mcmc_iter:
-            raise MlsynthConfigError(
-                "SPILLSYNTH: mcmc_burn must be smaller than mcmc_iter."
-            )
-        return values
-
-
-class DSCARConfig(BaseEstimatorConfig):
-    """Configuration for the Dynamic Synthetic Control AR (DSCAR) estimator.
-
-    Zheng & Chen (2024), JRSS-B 86(1):155-176, "Dynamic synthetic
-    control method for evaluating treatment effects in auto-regressive
-    processes." Extends Abadie-Diamond-Hainmueller (2010) to settings
-    with time-varying confounders, an auto-regressive outcome, spatial
-    dependence in the residuals, and possibly multiple treated units
-    that all turn on at a common intervention time.
-
-    The DSC weights are computed **per post-period** via empirical-
-    likelihood maximisation under per-period matching constraints
-    (equations 2.7-2.9 of the paper), allowing exact matches as
-    :math:`N_{co}, N_{tr} \\to \\infty` with :math:`T` fixed -- the
-    asymptotic regime that suits a typical air-pollution / hourly
-    panel.
-
-    Parameters
-    ----------
-    exog_covariates : list of str, optional
-        Time-varying exogenous covariate columns to include in the
-        per-period matching. ``None`` skips the covariate match (DSC
-        then matches on the lagged outcome only).
-    lagged_outcome : str, optional
-        Column name supplying the externally-computed
-        :math:`Y_{i, t-1}` value at the **first** sample period. For
-        later periods the lag is read off the panel itself. ``None``
-        drops the lag constraint at ``t = 1``.
-    placebo_reps : int
-        Number of normalised-placebo replications for the SE on the
-        DSC ATT (Section 3.2). ``0`` (default) skips placebo inference.
-    el_tolerance : float
-        Threshold for the QP residual ``mean|Z_1 - Z_0 w|`` at which
-        the EL refinement step is attempted; matches the R reference's
-        default ``0.01``. Smaller values fall back to QP weights more
-        often.
-    fdr_alpha : float
-        Significance level for the BY-adjusted pre-period
-        unconfoundedness test (Section 3.1).
-    seed : int
-        RNG seed for the placebo draw.
-
-    References
-    ----------
-    Zheng, X., & Chen, S. X. (2024). *Dynamic synthetic control method
-    for evaluating treatment effects in auto-regressive processes.*
-    Journal of the Royal Statistical Society Series B, 86(1):155-176.
-    """
-
-    exog_covariates: Optional[List[str]] = Field(
-        default=None,
-        description=(
-            "Time-varying exogenous covariate column names. None to "
-            "skip the covariate match (lagged-outcome match only)."
-        ),
-    )
-    lagged_outcome: Optional[str] = Field(
-        default=None,
-        description=(
-            "Column carrying Y_{i, t-1} at the first sample period. "
-            "Required to include the lag match at t = 1."
-        ),
-    )
-    placebo_reps: int = Field(
-        default=0,
-        description=(
-            "Number of normalised-placebo replications for the DSC "
-            "ATT standard error. 0 disables placebo inference."
-        ),
-    )
-    el_tolerance: float = Field(
-        default=1e-2,
-        description=(
-            "Mean-absolute-mismatch threshold for triggering the EL "
-            "refinement step. R reference default: 0.01."
-        ),
-    )
-    fdr_alpha: float = Field(
-        default=0.05,
-        description="Significance level for the BY pre-period test.",
-    )
-    seed: int = Field(default=0, description="RNG seed for placebo draws.")
 
 
 class PANGEOConfig(BaseModel):
@@ -1137,49 +484,6 @@ class PANGEOConfig(BaseModel):
         description="Plot treatment vs control aggregate trajectories per arm.")
     save: Union[bool, str] = Field(
         default=False, description="Save the plot (True or a filename base).")
-
-
-class PROXIMALConfig(BaseEstimatorConfig):
-    """Configuration for the Proximal Inference (PROXIMAL) estimator."""    # Override counterfactual_color from Base to match PROXIMAL's default and usage
-    counterfactual_color: Union[str, List[str]] = Field(default_factory=lambda: ["grey", "red", "blue"], description="Color(s) for counterfactual lines in plots. Can be a single color string or a list of color strings for multiple counterfactuals.")
-
-    methods: List[str] = Field(..., min_length=1, description='Which estimators to run. Any of "PI", "PIS", "PIPost", "SPSC", "DR", "PIPW". The estimator runs exactly these, and each method\'s required inputs must be present.')
-    donors: List[Union[str, int]] = Field(..., min_length=1, description="List of donor unit identifiers. Must not be empty.")
-    surrogates: List[Union[str, int]] = Field(default_factory=list, description="List of surrogate unit identifiers.")
-    vars: Dict[str, List[str]] = Field(default_factory=dict, description='Dictionary specifying proxy variables. Requires "donorproxies" when PI/PIS/PIPost is requested; requires "surrogatevars" when PIS/PIPost is requested. SPSC needs no proxies.')
-
-    spsc_detrend: bool = Field(default=True, description="Whether SPSC detrends the treated outcome against a B-spline time trend (SPSC-DT vs SPSC-NoDT).")
-    spsc_lambda: Optional[float] = Field(default=None, description="log10 ridge penalty for SPSC. If None, selected by leave-one-out cross-validation.")
-    spsc_spline_df: int = Field(default=5, ge=3, description="Degrees of freedom of the SPSC detrend B-spline basis.")
-    spsc_conformal: bool = Field(default=False, description="Whether to compute SPSC conformal prediction intervals for the per-period treatment effect.")
-    spsc_conformal_periods: Optional[List[int]] = Field(default=None, description="Absolute post-period indices to cover with SPSC conformal intervals. If None, every post-treatment period is covered.")
-
-    @model_validator(mode='after')
-    def check_methods_and_vars(cls, values: Any) -> Any:
-        valid_methods = {"PI", "PIS", "PIPost", "SPSC", "DR", "PIPW"}
-        methods = list(values.methods)
-        unknown = [m for m in methods if m not in valid_methods]
-        if unknown:
-            raise MlsynthConfigError(
-                f"Unknown PROXIMAL method(s) {unknown}. Valid choices: 'PI', 'PIS', 'PIPost', 'SPSC', 'DR', 'PIPW'."
-            )
-
-        vars_dict = values.vars
-        needs_donorproxies = any(m in methods for m in ("PI", "PIS", "PIPost", "DR", "PIPW"))
-        needs_surrogates = any(m in methods for m in ("PIS", "PIPost"))
-
-        if needs_donorproxies and not (isinstance(vars_dict.get("donorproxies"), list) and vars_dict.get("donorproxies")):
-            raise MlsynthConfigError(
-                "Config 'vars' must contain a non-empty list for 'donorproxies' when PI/PIS/PIPost is requested."
-            )
-        if needs_surrogates:
-            if not values.surrogates:
-                raise MlsynthConfigError("PIS/PIPost require a non-empty 'surrogates' list.")
-            if not (isinstance(vars_dict.get("surrogatevars"), list) and vars_dict.get("surrogatevars")):
-                raise MlsynthConfigError(
-                    "Config 'vars' must contain a non-empty list for 'surrogatevars' when PIS/PIPost is requested."
-                )
-        return values
 
 
 class FSCMConfig(BaseEstimatorConfig):
@@ -1889,139 +1193,6 @@ class SPCDConfig(BaseMAREXConfig):
 
         return values
 
-
-class SCMOConfig(BaseEstimatorConfig):
-    """Configuration for the Synthetic Control with Multiple Outcomes (SCMO) estimator."""
-
-    addout: Union[str, List[str]] = Field(
-        default_factory=list,
-        description="Auxiliary outcome variable(s) for outcome stacking. Used to build the matching spec when `spec` is not given.",
-    )
-    method: str = Field(
-        default="TLP",
-        description="Legacy method selector: 'TLP', 'SBMF', or 'BOTH'. Maps to schemes when `schemes` is not given.",
-        pattern="^(TLP|SBMF|BOTH)$",
-    )
-    conformal_alpha: float = Field(
-        default=0.1,
-        description="Miscoverage rate for conformal prediction intervals (e.g., 0.1 for 90% CI).",
-        gt=0, lt=1,
-    )
-    spec: Optional[Dict[str, Any]] = Field(
-        default=None,
-        description="Spec-driven matching matrix: {'year': int|list[int], 'vars': {name: column | (column, op)}}, op in {level,log,per_capita,raw}. If None, built from outcome+addout over the pre-period.",
-    )
-    schemes: Optional[List[str]] = Field(
-        default=None,
-        description="Weighting schemes to run: any of 'concatenated','averaged','separate','MA'. If None, derived from `method` (TLP->concatenated, SBMF->averaged, BOTH->[concatenated,averaged,MA]).",
-    )
-    demean: bool = Field(
-        default=False,
-        description="Intercept-shift the counterfactual (Doudchenko-Imbens / Sun-Ben-Michael-Feller level adjustment).",
-    )
-    conformal_q: float = Field(
-        default=1.0,
-        description="Norm exponent q of the CWZ conformal test statistic S_q (1 = average effect; larger targets sparse/large effects across outcomes).",
-        gt=0,
-    )
-
-
-
-
-
-class SIConfig(BaseEstimatorConfig):
-    """Configuration for the Synthetic Interventions (SI) estimator.
-
-    Implements:
-
-        Agarwal, A., Shah, D., & Shen, D. (2026). "Synthetic Interventions:
-        Extending Synthetic Controls to Multiple Treatments." Operations
-        Research 74(2):840-859.
-
-    SI estimates the focal target unit's counterfactual outcome under each
-    alternative intervention in ``inters`` via SI-PCR: regress the target's
-    pre-treatment control outcomes onto the rank-``k`` denoised donor pool, then
-    apply the weights to the donor pool's post-intervention outcomes.
-
-    Parameters
-    ----------
-    inters : list of str
-        Binary indicator columns naming the alternative interventions; for
-        each, the units flagged ``1`` form that intervention's donor pool.
-    rank_method : {"donoho", "usvt", "cumvar", "fixed"}
-        Spectral-rank rule for the donor pre-matrix. ``"donoho"`` (default)
-        reproduces the paper's exact Gavish-Donoho optimal hard threshold
-        (evaluated at ``ratio = T0 / Nd``); ``"usvt"`` is the same threshold at
-        the canonical ``min/max`` aspect ratio; ``"cumvar"`` keeps enough
-        components for ``cumvar_threshold`` of the spectral energy; ``"fixed"``
-        uses ``rank``.
-    rank : int or None
-        Explicit spectral rank ``k`` for ``rank_method="fixed"``.
-    cumvar_threshold : float
-        Cumulative-energy target in ``(0, 1]`` for ``rank_method="cumvar"``.
-    bias_correct : bool
-        Use the bias-corrected SI-PCR estimator (default ``True``), which fits
-        weights on a rank-complete donor subset and enables asymptotic-normality
-        intervals (Section 4.3). ``False`` gives plain SI-PCR (eq. 10), point
-        estimate only.
-    variance : {"double", "units", "time_iv"}
-        Noise-variance estimator behind the interval. ``"double"`` (default)
-        matches the paper's code (a d.o.f.-weighted combination); ``"units"`` is
-        the main-text eq. 14; ``"time_iv"`` uses the donor post-period residual.
-    interval : {"confidence", "prediction"}
-        Interval type. ``"confidence"`` (default) is the eq.-13 CI for the
-        counterfactual mean; ``"prediction"`` is the wider interval the paper's
-        case study uses for coverage validation.
-    alpha : float
-        Two-sided significance level for the intervals.
-    display_graphs : bool
-        Show the observed-vs-counterfactual plot after fitting.
-    """
-
-    inters: List[str] = Field(
-        ..., min_length=1,
-        description="Binary indicator columns naming the alternative "
-                    "interventions (donor pool = units flagged 1).",
-    )
-    rank_method: Literal["donoho", "usvt", "cumvar", "fixed"] = Field(
-        default="donoho",
-        description="Spectral-rank rule: paper's Gavish-Donoho ('donoho'), "
-                    "min/max Gavish-Donoho ('usvt'), cumulative-energy "
-                    "('cumvar'), or explicit ('fixed').",
-    )
-    rank: Optional[int] = Field(
-        default=None, ge=1,
-        description="Explicit spectral rank k for rank_method='fixed'.",
-    )
-    cumvar_threshold: float = Field(
-        default=0.95, gt=0.0, le=1.0,
-        description="Cumulative-energy target for rank_method='cumvar'.",
-    )
-    bias_correct: bool = Field(
-        default=True,
-        description="Use the bias-corrected SI-PCR estimator (enables intervals).",
-    )
-    variance: Literal["double", "units", "time_iv"] = Field(
-        default="double",
-        description="Noise-variance estimator: paper's weighted ('double'), "
-                    "main-text eq. 14 ('units'), or post-period ('time_iv').",
-    )
-    interval: Literal["confidence", "prediction"] = Field(
-        default="confidence",
-        description="Interval type: confidence (eq. 13) or prediction.",
-    )
-    alpha: float = Field(
-        default=0.05, gt=0.0, lt=1.0,
-        description="Two-sided significance level for the intervals.",
-    )
-
-    @model_validator(mode="after")
-    def _check_si_params(cls, values: Any) -> Any:
-        if values.rank_method == "fixed" and values.rank is None:
-            raise MlsynthConfigError(
-                "rank_method='fixed' requires an explicit positive `rank`."
-            )
-        return values
 
 class NSCConfig(BaseEstimatorConfig):
     """Configuration for the Nonlinear Synthetic Control (NSC) estimator.
@@ -3442,6 +2613,16 @@ class DesignResult(MlsynthResult):
 # BaseEstimatorConfig from here).
 # ---------------------------------------------------------------------------
 _RELOCATED_CONFIGS = {
+    "RMSIConfig": "mlsynth.utils.rmsi_helpers.config",
+    "SPOTSYNTHConfig": "mlsynth.utils.spotsynth_helpers.config",
+    "SNNConfig": "mlsynth.utils.snn_helpers.config",
+    "CTSCConfig": "mlsynth.utils.ctsc_helpers.config",
+    "ISCMConfig": "mlsynth.utils.iscm_helpers.config",
+    "SPILLSYNTHConfig": "mlsynth.utils.spillsynth_helpers.config",
+    "DSCARConfig": "mlsynth.utils.dscar_helpers.config",
+    "PROXIMALConfig": "mlsynth.utils.proximal_helpers.config",
+    "SCMOConfig": "mlsynth.utils.scmo_helpers.config",
+    "SIConfig": "mlsynth.utils.si_helpers.config",
     "FMAConfig": "mlsynth.utils.fma_helpers.config",
     "PDAConfig": "mlsynth.utils.pda_helpers.config",
     "CLUSTERSCConfig": "mlsynth.utils.clustersc_helpers.config",
